Moving Platform
To set up the moving platform, first drag it into a level scene.

Create a curve for the path 2D (this is the path the platform will follow). you can also change the height and length of the platform as well as the speed.

If you want the platform to move back and forth Then set "Loop" to false. and change the "speed scale" variable to set how fast you want it to move.

If you want it to reach the end of the path and immediately respawn at the beginning then turn Loop on and modify the "Speed" variable to set how fast it moves.