Rotating Platform
To set up the rotating platform, just drag it into a level scene, change the height and radius to your liking and set the speed scale to how fast you want it to rotate.

Right now it makes a full rotation every 4 seconds, so setting the speed scale to 0.5 will make it 8 seconds and setting it to 2 will make it 2 seconds.