New levels (laser acting weird)
I created a new floor asses so we can have a standardized asset to use as a wall and floor. i put that within the assets folder.

I resized the crate so the physics would be more responsive and not be wonky due to the scale being off.

and I made two levels, inside a new level folder. The second level is the one with the weird interaction when more than 1 laser is in a scene.